requirements.txt
beautifulsoup4==1.21.6
lxml==4.9.1
requests==2.28.1
regex==2022.9.13
numpy==1.23.3
python-dotenv==0.21.0